add remote plugin discovery test case
